requirements.txt
ansicon==1.89.0
asttokens==2.4.0
backcall==0.2.0
blessed==1.20.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
editor==1.6.6
executing==1.2.0
filelock==3.13.1
fonttools==4.42.1
inquirer==3.2.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jinxed==1.2.1
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
missingno==0.5.2
nest-asyncio==1.5.8
numpy==1.25.2
packaging==23.1
pandas==2.1.1
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
plotly-express==0.4.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyaml==23.9.6
pydobot @ git+https://github.com/luismesas/pydobot.git@2ff8d7e1318ac9b7bc32bd33ff327d343f44927c
Pygments==2.16.1
pyparsing==3.1.1
pyserial==3.4
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
readchar==4.0.5
rich==13.7.0
runs==1.2.2
scikit-learn==1.3.0
scikit-optimize==0.9.0
scipy==1.11.2
seaborn==0.12.2
shellingham==1.5.4
six==1.16.0
stack-data==0.6.2
statsmodels==0.14.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.10.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
virtualenv==20.25.0
wcwidth==0.2.7
xmod==1.8.1
yaspin==3.0.1
